+++ changelog.md
@@ -3,6 +3,7 @@ The latest version may not be released!
 
 ## 3.1.1
 - Fix for ESP32 send Error, removed `USE_SOFT_SEND_PWM` macro.
+- Added Onkyo protocol.
 
 ## 3.1.0
 - Generation of PWM by software is active by default.